tests/wpt/tests/html/dom/render-blocking/element-render-blocking-035.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/utils.js"></script>
<title>link rel=expect: base URL should be OK with relative URLs</title>

<base href="dummy.html">
<link rel=expect href="../render-blocking/element-render-blocking-035.html#second" blocking="render">
<script>
async_test((t) => {
  requestAnimationFrame(() => {
    t.step(() => assert_true(!!document.getElementById("second")));
    t.step(() => assert_false(!!document.getElementById("last")), "the second element should already unblock rendering");
    t.done();
  });
}, "relative URLs that match this document are OK, regarless of <base>");
</script>
</head>
<body>
  <div id="first"></div>
  <script>
          generateParserDelay();
  </script>
  <div id="second">
  </div>
  <script>
          generateParserDelay();
  </script>
  <div id="last"></div>
</body>
